requirements.txt
pytest==7.2.2
Appium-Python-Client==2.8.1
